audio_policy: Add extra check before restoring device in stopOutput
When certain game apks run with sound enabled, UI lag will be
observed.

This is due to device to be restored with double HAL latency even
when it's the same with the previous device, since setOutputDevice
is synchronous call now, then it will block the subsequent calling
to APM until the latency time elapsed.

When the device to be restored is the same as previous device,
set the latency to 0.

Change-Id: I2c92675cedd8930430f81147c7f733d8dd70798c
CRs-Fixed: 886459